module09_asynchronous_programming/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
annotated-types==0.7.0
anyio==3.7.1
appnope==0.1.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
backoff==2.2.1
bullet==2.2.0
catboost==1.2.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
chroma-hnswlib==0.7.1
chroma-migrate==0.0.2
chromadb==0.4.0
click==8.1.5
clickhouse-connect==0.6.6
colorama==0.4.6
coloredlogs==15.0.1
comm==0.1.3
contourpy==1.1.1
cryptography==41.0.3
cycler==0.11.0
dataclasses-json==0.5.12
debugpy==1.6.6
decorator==5.1.1
distlib==0.3.7
docopt==0.6.2
duckdb==0.7.1
elastic-transport==8.17.0
elasticsearch==8.17.0
executing==1.2.0
fastapi==0.99.1
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.42.1
frozenlist==1.3.3
fsspec==2024.12.0
graphviz==0.20.1
h11==0.14.0
httpcore==1.0.4
httptools==0.6.0
httpx==0.25.2
huggingface-hub==0.31.2
humanfriendly==10.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.0
ipykernel==6.22.0
ipython==8.12.0
jedi==0.18.2
Jinja2==3.1.5
joblib==1.3.2
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.1.0
jupyter_core==5.3.0
kiwisolver==1.4.5
langchain==0.0.235
langchain-core==0.3.60
langchain-elasticsearch==0.3.0
langchain-huggingface==0.2.0
langsmith==0.3.42
lz4==4.3.2
MarkupSafe==3.0.2
marshmallow==3.19.0
matplotlib==3.8.0
matplotlib-inline==0.1.6
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.6
networkx==3.4.2
numexpr==2.8.4
numpy==1.25.1
ollama==0.1.6
onnxruntime==1.15.1
openai==0.27.2
openapi-schema-pydantic==1.2.4
orjson==3.10.14
overrides==7.3.1
packaging==24.2
pandas==2.0.3
parso==0.8.3
patsy==0.5.3
pdfminer.six==20220524
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
platformdirs==3.11.0
plotly==5.17.0
posthog==3.0.1
prompt-toolkit==3.0.38
protobuf==4.23.4
psutil==5.9.4
ptyprocess==0.7.0
pulsar-client==3.2.0
pure-eval==0.2.2
py-pdf-parser==0.11.0
pycparser==2.21
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.14.0
pyparsing==3.1.1
pypdf==3.12.2
PyPika==0.48.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
pyzmq==25.0.2
regex==2023.6.3
requests==2.28.2
requests-toolbelt==1.0.0
safetensors==0.5.2
scikit-learn==1.3.1
scikit-posthocs==0.7.0
scipy==1.11.3
seaborn==0.13.0
sentence-transformers==3.3.1
simsimd==6.2.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.19
stack-data==0.6.2
starlette==0.27.0
statsmodels==0.14.0
sympy==1.13.1
tenacity==8.2.2
threadpoolctl==3.2.0
tiktoken==0.4.0
tokenizers==0.21.0
torch==2.5.1
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.48.0
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2023.3
urllib3==1.26.15
uvicorn==0.23.1
uvloop==0.17.0
virtualenv==20.24.6
Wand==0.6.10
watchfiles==0.19.0
wcwidth==0.2.6
websockets==11.0.3
xgboost==2.0.0
yarl==1.8.2
zipp==3.16.2
zstandard==0.23.0
